iOS: stream cloud dictation through Writingmate realtime endpoints
Share the macOS OpenAI realtime client, client_secrets minting, and
finish-gate types in WhisperMateShared. Cloud-mode iOS capture (sheet,
inline, and keyboard host) now streams 24 kHz PCM during recording and
prefers the realtime transcript on stop, with the existing batch upload
path as fallback. Offline/Parakeet and meetings/diarization stay on
batch. The durable AVAudioRecorder file path is unchanged.
